BAD/src/ticket_bot/requirements.txt
discord.py>=2.3.0
google-generativeai>=0.3.0
google-api-python-client>=2.100.0
python-dotenv>=1.0.0
